+++ CHANGE_LOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Send `omnichannel-chat-widget` as part of `ocUserAgent`
+
+### Changed
+
+- Uptake [@microsoft/omnichannel-chat-sdk@1.8.3](https://www.npmjs.com/package/@microsoft/omnichannel-chat-sdk/v/1.8.3)
+
 ## [1.6.5] 2024-05-10
 
 ### Added